+++ src/components/BookCover.tsx
@@ -4,6 +4,7 @@ import * as React from "react";
 import { AspectRatio } from "@theme-ui/components";
 import { MediumIcon } from "./MediumIndicator";
 import { AnyBook } from "interfaces";
+import LazyImage from "components/LazyImage";
 
 /**
  * This is meant to be a book cover. Primarily the image and styling,
@@ -51,7 +52,7 @@ const BookCover: React.FC<{
       >
         <MediumIcon book={book} sx={{ height: "30%", fill: "ui.gray.dark" }} />
       </AspectRatio>
-      <img
+      <LazyImage
         alt={`Cover of book: ${book.title}`}
         src={imageUrl}
         onError={handleError}